requirements.txt
apiview==1.3.25
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-import-export==2.0.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
pipenv==2018.11.26
pylint==2.4.4
pytz==2019.3
PyYAML==5.3
reversion==0.2
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
toml==0.10.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
